requirements.txt
accelerate==0.26.1
adjustText==1.0.4
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
datasets==2.16.1
deepspeed==0.12.6
Deprecated==1.2.14
einops==0.7.0
faiss-cpu==1.7.4
fastapi==0.110.0
fasttext==0.9.2
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.28.1
googleapis-common-protos==1.62.0
huggingface-hub==0.20.2
ipdb==0.13.13
info-nce-pytorch==0.1.4
jsonlines==4.0.0
mteb==1.1.1
ninja==1.11.1.1
nltk==3.8.1
numpy==1.23.0
omegaconf==2.3.0
pandas==2.0.3
peft==0.7.1
Pillow==9.3.0
PyYAML==6.0.1
rich==13.7.0
scikit-learn==1.3.2
scipy==1.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
smmap==5.0.1
sniffio==1.3.1
tokenizers==0.15.0
tomli==2.0.1
tqdm==4.66.1
transformers==4.36.2
uvicorn==0.27.1
wandb==0.16.2
openai>=1.0.0
submitit>=1.5.0
